+++ src/Repositories/TokenResetRepository.php
@@ -5,11 +5,9 @@ declare(strict_types=1);
 namespace TheBachtiarz\Auth\Repositories;
 
 use Exception;
-use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
-use Illuminate\Database\Query\Builder as QueryBuilder;
 use TheBachtiarz\Auth\Interfaces\Models\TokenResetInterface;
 use TheBachtiarz\Auth\Models\TokenReset;
 use TheBachtiarz\Base\App\Repositories\AbstractRepository;
@@ -37,10 +35,12 @@ class TokenResetRepository extends AbstractRepository
      */
     public function getByToken(string $token): TokenResetInterface
     {
-        $entity = TokenReset::getByToken($token)->first();
+        $this->modelBuilder(modelBuilder: TokenReset::getByToken($token));
+
+        $entity = $this->modelBuilder()->first();
         assert($entity instanceof TokenResetInterface);
 
-        if (! $entity) {
+        if (! $entity && $this->throwIfNullEntity()) {
             throw new ModelNotFoundException(sprintf("Token reset with token '%s' not found", $token));
         }
 
@@ -54,14 +54,13 @@ class TokenResetRepository extends AbstractRepository
      */
     public function getByIdentifier(string $identifier): Collection
     {
-        $builder = TokenReset::getByIdentifier($identifier);
-        assert($builder instanceof EloquentBuilder || $builder instanceof QueryBuilder);
+        $this->modelBuilder(modelBuilder: TokenReset::getByIdentifier($identifier));
 
-        if (! $builder->count()) {
+        if (! $this->modelBuilder()->count() && $this->throwIfNullEntity()) {
             throw new ModelNotFoundException(sprintf("Cannot find any token reset with identifier '%s'", $identifier));
         }
 
-        return $builder->get();
+        return $this->modelBuilder()->get();
     }
 
     /**
@@ -101,6 +100,10 @@ class TokenResetRepository extends AbstractRepository
     public function deleteByToken(string $token): bool
     {
         $tokenReset = $this->getByToken($token);
+
+        if (! $tokenReset) {
+            throw new ModelNotFoundException('Failed to delete token reset');
+        }
 
         return $this->deleteById($tokenReset->getId());
     }
